CDRIVER-2228 decide if a command is retryable during part assembly
This consolidates most of the decision making for a command's retryable
eligibility. Bulk operations will need to track if they include one or
more multi-document writes (akin to collation).

While the spec prohibits a generic command helper from injecting
txnNumber and supporting retryable writes, we are allowed to do so for
the write-specific command helper(s).