CDRIVER-2254 fix hang with "canonicalizeHostname"
The GSSAPI / Kerberos option canonicalizeHostname allows the driver to
authenticate when hosts report different hostnames than what is used in
the Kerberos database. We had used it both with Cyrus-SASL and with
Microsoft's SSPI, and it had an infinite loop.

Fix the infinite loop, and ignore the setting with SSPI, since SSPI
canonicalizes hostnames itself. The setting is only required for Cyrus.